8156 dbuf_evict_notify() does not need dbuf_evict_lock
Reviewed by: Dan Kimmel <dan.kimmel@delphix.com>
Reviewed by: Paul Dagnelie <pcd@delphix.com>

dbuf_evict_notify() holds the dbuf_evict_lock while checking if it should do the eviction itself
(because the evict thread is not able to keep up). This can result in massive lock contention.
It isn't necessary to hold the lock, because if we make the wrong choice occasionally, nothing
bad will happen. This commit results in a ~60% performance improvement for ARC-cached sequential reads.

Upstream bugs: DLPX-50674