WIP: Add new "mdb_kdump" module files
This commit introduces a new "mdb_kdump" file that's intended to be used
when running MDB against a Linux kernel compressed dump. Currently, the
only functionality that is included is a function that can read the
header information from the Linux dump file; the logic for this is
contained in the new kdump_open() function.

If this change is used to run MDB against a linux kernel dump, it will
fail because the kdump_sym_io() function hasn't been implemented.

For example, here's what it'd look like:

    $ mdb vmlinux-4.10.0-37-generic 201710251948/dump.201710251948
    mdb: failed to initialize target: No such file or directory