WIP: Add 0x10800000 to the value of ELF symbols
I have no idea why this patch works, but it seems to be necessary in
order for symbol to address translation to work. With this in place, I
can now use the "kmem_cache_alloc" symbol to dump that functions
instructions. For example:

    $ mdb vmlinux-4.10.0-37-generic 201710251948/dump.201710251948
    mdb: failed to read panicbuf and panic_reg -- current register set will be unavailable
    > kmem_cache_alloc::dis
    0xffffffff91a1cec0:             nopl   0x0(%rax,%rax)
    0xffffffff91a1cec5:             pushq  %rbp
    0xffffffff91a1cec6:             movq   %rsp,%rbp
    0xffffffff91a1cec9:             pushq  %r15
    0xffffffff91a1cecb:             pushq  %r14
    0xffffffff91a1cecd:             pushq  %r13
    0xffffffff91a1cecf:             pushq  %r12
    0xffffffff91a1ced1:             movq   %rdi,%r15
    0xffffffff91a1ced4:             pushq  %rbx
    0xffffffff91a1ced5:             movl   %esi,%r12d
    0xffffffff91a1ced8:             subq   $0x8,%rsp

Previously, using "kmem_cache_alloc" would not have returned the correct
address, and thus the disassembly output would not have been correct.

Obviously this patch is a hack, and probably doesn't work for all cases
(and/or maybe only works in this one particular case), but since I don't
understand what's going on, it's hard for me to know how to do anything
better.