WIP: Don't use hardcoded KERNELOFFSET value
This change removes the reliance on the 0x10800000ULL value being
hardcoded into the ELF functions, and instead reads in the value of
KERNELOFFSET from the dump's vmcoreinfo section, and uses this. After
doing some digging, I found out that the value previously used was due
to KASLR being enabled on the kernel that crash, which adds an offset to
the symbol addresses. This patch adds logic to read in this offset from
the dump file, and then plumb this value into the functions that needed
it when parsing the ELF vmlinux file.

A result of this change, is "::dis" now works correctly:

    > kmem_cache_alloc::dis ! head -n 20
    kmem_cache_alloc:               nopl   0x0(%rax,%rax)
    kmem_cache_alloc+5:             pushq  %rbp
    kmem_cache_alloc+6:             movq   %rsp,%rbp
    kmem_cache_alloc+9:             pushq  %r15
    kmem_cache_alloc+0xb:           pushq  %r14
    kmem_cache_alloc+0xd:           pushq  %r13
    kmem_cache_alloc+0xf:           pushq  %r12
    kmem_cache_alloc+0x11:          movq   %rdi,%r15
    kmem_cache_alloc+0x14:          pushq  %rbx
    kmem_cache_alloc+0x15:          movl   %esi,%r12d
    kmem_cache_alloc+0x18:          subq   $0x8,%rsp
    kmem_cache_alloc+0x1c:          movl   +0xd7976a(%rip),%ebx     <gfp_allowed_mask>
    kmem_cache_alloc+0x22:          movq   0x8(%rbp),%r13
    kmem_cache_alloc+0x26:          andl   %esi,%ebx
    kmem_cache_alloc+0x28:          testl  $0x400000,%ebx
    kmem_cache_alloc+0x2e:          jne    +0x10d   <kmem_cache_alloc+0x141>
    kmem_cache_alloc+0x34:          jmp    +0x70    <kmem_cache_alloc+0xa9>
    kmem_cache_alloc+0x39:          movq   %r15,%r14
    kmem_cache_alloc+0x3c:          testq  %r14,%r14
    kmem_cache_alloc+0x3f:          je     +0x145   <kmem_cache_alloc+0x18a>

Previously the function offsets would not have been properly displayed.
Additionally, only the instructions for the beginning of the function
would have been shown, whereas now, the entire function is shown.