WIP: "lkd" library and mdb target for Linux Kdump
This patch adds a new "lkd" library whose purpose is to be used to read
and interpret a Linux kernel Kdump file. Additionally, a new "lkd" mdb
target has been added to make use of this new library, and enable mdb to
debug a Linux kernel Kdump file.

Both the new library and mdb target only have limited functionality, but
they will hopefully provide a foundation to build on.