WIP: Add initial plumbing for lkd target dcmds
This simple change adds the initial plumbing to start implementating
dcmds for the new lkd target. So far, this adds the "::regs" and "$r"
dcmds that are used for dumping the registers, but doesn't contain the
actual implementation for gathering and dumping the registers.

Here's an example before this change:

    $ mdb -l vmlinux-4.10.0-37-generic serapheim.dump
    > ::regs
    mdb: command is not supported by current target
    > $r
    mdb: command is not supported by current target

And here that same example with this change:

   $ mdb -l vmlinux-4.10.0-37-generic serapheim.dump
   > ::regs
   > $r

Ta da!